Add comments to SQL migration
