+++ src/main/resources/config/liquibase/changelog/20230320100000_changelog.xml
@@ -14,32 +14,42 @@
                                  initiallyDeferred="false" referencedColumnNames="id"
                                  referencedTableName="programming_exercise_test_case" validate="true"/>
         <sql>
-            WITH tp AS (SELECT e.id, e.template_participation_id FROM exercise e
-                        WHERE e.discriminator = 'P' AND e.template_participation_id IS NOT NULL)
-               , sp AS (SELECT e.id, e.solution_participation_id FROM exercise e
-                        WHERE e.discriminator = 'P' AND e.solution_participation_id IS NOT NULL)
-
+            <!-- Helper tables with all template and solution participation ids and their corresponding exercises
+            needed since participation.exerciseId is null for these types of participation -->
+            WITH tp AS (SELECT e.id, e.template_participation_id
+                        FROM exercise e
+                        WHERE e.discriminator = 'P'
+                          AND e.template_participation_id IS NOT NULL)
+               , sp AS (SELECT e.id, e.solution_participation_id
+                        FROM exercise e
+                        WHERE e.discriminator = 'P'
+                          AND e.solution_participation_id IS NOT NULL)
+               <!-- Maps every feedback to their exercise id -->
                , fe AS (SELECT f.id, f.text, e.id AS exercise_id
                         FROM Feedback f
-                            LEFT JOIN result r ON f.result_id = r.id
-                            LEFT JOIN participation p ON r.participation_id = p.id
-                           , exercise e
+                          LEFT JOIN result r ON f.result_id = r.id
+                          LEFT JOIN participation p ON r.participation_id = p.id
+                        , exercise e
                         WHERE f.type = 3
-                            AND f.text IS NOT NULL
-                            AND f.reference IS NULL
-                            AND f.text NOT LIKE 'SubPolFeedbackIdentifier%'
-                            AND (p.exercise_id = e.id
-                                OR p.id = (SELECT tp.template_participation_id FROM tp WHERE tp.id = e.id)
-                                OR p.id = (SELECT sp.solution_participation_id FROM sp WHERE sp.id = e.id)))
+                          AND f.text IS NOT NULL
+                          AND f.reference IS NULL
+                          AND f.text NOT LIKE 'SubPolFeedbackIdentifier%'
+                          AND (p.exercise_id = e.id
+                            OR p.id = (SELECT tp.template_participation_id FROM tp WHERE tp.id = e.id)
+                            OR p.id = (SELECT sp.solution_participation_id FROM sp WHERE sp.id = e.id)))
 
             UPDATE feedback f
-            SET test_case_id = (
-                SELECT tc.id
-                FROM programming_exercise_test_case tc
-                    LEFT JOIN fe ON fe.id = f.id
-                WHERE tc.test_name = fe.text
-                    AND tc.exercise_id = fe.exercise_id
-            )
+            <!-- Find the testcase by matching the text / test name and the exercise -->
+            SET f.test_case_id = (SELECT tc.id
+                                FROM programming_exercise_test_case tc,
+                                     fe
+                                WHERE tc.test_name = fe.text
+                                  AND tc.exercise_id = fe.exercise_id
+                                  AND fe.id = f.id)
+            WHERE f.type = 3 <!-- Automatic -->
+                AND f.text IS NOT NULL
+                AND f.text NOT LIKE 'SubPolFeedbackIdentifier%' <!-- Filter out Submission policy feedback -->
+                AND f.reference IS NULL <!-- Filter out SCA feedback -->
         </sql>
 
         <sql>
